+++ src/engine/Graphics/Text.ts
@@ -4,15 +4,31 @@ import { SpriteFont } from './SpriteFont';
 import { Graphic, GraphicOptions } from './Graphic';
 import { Color } from '../Color';
 import { Font } from './Font';
-import { watch } from '../Util/Watch';
 
 export interface TextOptions {
+  /**
+   * Text to draw
+   */
   text: string;
+
+  /**
+   * Optionally override the font color, currently unsupported by SpriteFont
+   */
   color?: Color;
+
+  /**
+   * Optionally specify a font, if none specified a default font is used (System sans-serif 10 pixel)
+   */
   font?: Font | SpriteFont;
 }
 
+/**
+ * Represent Text graphics in excalibur
+ *
+ * Useful for in game labels, ui, or overlays
+ */
 export class Text extends Graphic {
+  public color?: Color;
   constructor(options: TextOptions & GraphicOptions) {
     super(options);
     // This order is important font, color, then text
@@ -24,7 +40,7 @@ export class Text extends Graphic {
   public clone(): Text {
     return new Text({
       text: this.text.slice(),
-      color: this.color.clone(),
+      color: this.color?.clone() ?? Color.Black,
       font: this.font.clone()
     });
   }
@@ -36,21 +52,9 @@ export class Text extends Graphic {
 
   public set text(value: string) {
     this._text = value;
-    this.font.updateText(value);
-  }
-
-  // TODO SpriteFont doesn't support a color yet :(
-  public get color() {
-    if (this.font instanceof Font) {
-      return this.font.color;
-    }
-    return Color.Black;
-  }
-
-  public set color(color: Color) {
-    if (this.font instanceof Font) {
-      this.font.color = color;
-    }
+    const bounds = this.font.measureText(this._text);
+    this._textWidth = bounds.width;
+    this._textHeight = bounds.height;
   }
 
   private _font: Font | SpriteFont;
@@ -58,38 +62,50 @@ export class Text extends Graphic {
     return this._font;
   }
   public set font(font: Font | SpriteFont) {
-    if (font instanceof Font) {
-      this._font = watch(font, (font) => font.flagDirty());
-    } else {
-      this._font = font;
-    }
+    this._font = font;
   }
+
+  private _textWidth: number = 0;
 
   public get width() {
-    return this.font.width;
+    if (this._textWidth === 0) {
+      this._calculateDimension();
+    }
+    return this._textWidth;
   }
 
+  private _textHeight: number = 0;
   public get height() {
-    return this.font.height;
+    if (this._textHeight === 0) {
+      this._calculateDimension();
+    }
+    return this._textHeight;
+  }
+
+  private _calculateDimension() {
+    const { width, height } = this.font.measureText(this._text);
+    this._textWidth = width;
+    this._textHeight = height;
   }
 
   public get localBounds(): BoundingBox {
-    return this.font.localBounds;
+    return this.font.measureText(this._text);
   }
 
-  protected _rotate(_ex: ExcaliburGraphicsContext) {
+  protected override _rotate(_ex: ExcaliburGraphicsContext) {
     // None this is delegated to font
     // This override erases the default behavior
   }
 
-  protected _flip(_ex: ExcaliburGraphicsContext) {
+  protected override _flip(_ex: ExcaliburGraphicsContext) {
     // None this is delegated to font
     // This override erases the default behavior
   }
 
-  protected _drawImage(ex: ExcaliburGraphicsContext, x: number, y: number) {
+  protected override _drawImage(ex: ExcaliburGraphicsContext, x: number, y: number) {
+    let color = Color.Black;
     if (this.font instanceof Font) {
-      this.font.color = this.color;
+      color = this.color ?? this.font.color;
     }
     this.font.flipHorizontal = this.flipHorizontal;
     this.font.flipVertical = this.flipVertical;
@@ -97,6 +113,14 @@ export class Text extends Graphic {
     this.font.rotation = this.rotation;
     this.font.origin = this.origin;
     this.font.opacity = this.opacity;
-    this.font.render(ex, this._text, x, y);
+
+    const { width, height } = this.font.measureText(this._text);
+    this._textWidth = width;
+    this._textHeight = height;
+
+    this.font.render(ex, this._text, color, x, y);
+    if (this.font.showDebug) {
+      ex.debug.drawRect(x - width, y - height, width * 2, height * 2);
+    }
   }
 }